+++ apps/hyperdrive-trading/src/ui/hyperdrive/hooks/usePrepareSharesIn.ts
@@ -1,9 +1,9 @@
-import { ReadHyperdrive } from "@delvtech/hyperdrive-viem";
 import {
   AppConfig,
   appConfig,
   findHyperdriveConfig,
-} from "@hyperdrive/appconfig";
+} from "@delvtech/hyperdrive-appconfig";
+import { ReadHyperdrive } from "@delvtech/hyperdrive-viem";
 import { useQuery } from "@tanstack/react-query";
 import { makeQueryKey } from "src/base/makeQueryKey";
 import { QueryStatusWithIdle, getStatus } from "src/base/queryStatus";