+++ apps/hyperdrive-trading/src/ui/hyperdrive/shorts/OpenShortPreview/OpenShortPreview.tsx
@@ -1,13 +1,13 @@
 import { parseFixed } from "@delvtech/fixed-point-wasm";
-import { ArrowRightIcon } from "@heroicons/react/16/solid";
-import { ChevronDownIcon } from "@heroicons/react/20/solid";
-import { ClockIcon } from "@heroicons/react/24/outline";
 import {
   HyperdriveConfig,
   TokenConfig,
   appConfig,
   findBaseToken,
-} from "@hyperdrive/appconfig";
+} from "@delvtech/hyperdrive-appconfig";
+import { ArrowRightIcon } from "@heroicons/react/16/solid";
+import { ChevronDownIcon } from "@heroicons/react/20/solid";
+import { ClockIcon } from "@heroicons/react/24/outline";
 import classNames from "classnames";
 import { ReactElement, useState } from "react";
 import Skeleton from "react-loading-skeleton";