+++ packages/fixed-point-wasm/fixed_point_wasm.d.ts
@@ -439,10 +439,8 @@ export interface InitOutput {
   readonly fixedpoint_sub: (a: number, b: number, c: number, d: number) => void;
   readonly fixedpoint_mulDivDown: (a: number, b: number, c: number, d: number, e: number) => void;
   readonly fixedpoint_mulDivUp: (a: number, b: number, c: number, d: number, e: number) => void;
-  readonly fixedpoint_mulDown: (a: number, b: number, c: number, d: number) => void;
   readonly fixedpoint_mulUp: (a: number, b: number, c: number, d: number) => void;
   readonly fixedpoint_mul: (a: number, b: number, c: number, d: number) => void;
-  readonly fixedpoint_divDown: (a: number, b: number, c: number, d: number) => void;
   readonly fixedpoint_divUp: (a: number, b: number, c: number, d: number) => void;
   readonly fixedpoint_div: (a: number, b: number, c: number, d: number) => void;
   readonly fixedpoint_pow: (a: number, b: number, c: number, d: number) => void;
@@ -461,6 +459,8 @@ export interface InitOutput {
   readonly fixedpoint_toFixed: (a: number, b: number) => number;
   readonly initialize: () => void;
   readonly fixedpoint_valueOf: (a: number, b: number) => void;
+  readonly fixedpoint_mulDown: (a: number, b: number, c: number, d: number) => void;
+  readonly fixedpoint_divDown: (a: number, b: number, c: number, d: number) => void;
   readonly getVersion: (a: number) => void;
   readonly fixed: (a: number, b: number, c: number) => void;
   readonly parseFixed: (a: number, b: number, c: number) => void;